--- cg/meta/deliver/__init__.py
@@ -1,2 +1,0 @@
-from .deliver_ticket import DeliverTicketAPI
-from .deliver import DeliverAPI
