+++ cg/meta/upload/nf_analysis.py
@@ -6,7 +6,7 @@ import logging
 import click
 
 from cg.cli.generate.report.base import generate_delivery_report
-from cg.cli.upload.clinical_delivery import upload_clinical_delivery
+
 from cg.cli.upload.scout import upload_to_scout
 from cg.constants import (
     REPORT_SUPPORTED_DATA_DELIVERY,
