+++ tests/services/fastq_file_service/test_fastq_file_service.py
@@ -1,7 +1,12 @@
 from pathlib import Path
+
+import pytest
+from cg.constants.constants import ReadDirection
+
 from cg.services.fastq_concatenation_service.fastq_concatenation_service import (
     FastqConcatenationService,
 )
+from cg.services.fastq_concatenation_service.utils import generate_concatenated_fastq_delivery_path
 
 
 def test_empty_directory(fastq_file_service: FastqConcatenationService, tmp_path):
@@ -104,3 +109,32 @@ def test_concatenate_missing_reverse(
 
     # THEN reverse reads should not exist
     assert not reverse_output_path.exists()
+
+
+@pytest.mark.parametrize(
+    "fastq_directory, sample_name, direction, expected_output_path",
+    [
+        (
+            Path("/path/to/fastq_directory"),
+            "sample1",
+            ReadDirection.FORWARD,
+            Path("/path/to/fastq_directory/sample1_1.fastq.gz"),
+        ),
+        (
+            Path("/path/to/fastq_directory"),
+            "sample1",
+            ReadDirection.REVERSE,
+            Path("/path/to/fastq_directory/sample1_2.fastq.gz"),
+        ),
+    ],
+    ids=["forward", "reverse"],
+)
+def test_generate_concatenated_fastq_delivery_path(
+    fastq_directory: Path, sample_name: str, direction: int, expected_output_path: Path
+):
+    assert (
+        generate_concatenated_fastq_delivery_path(
+            fastq_directory=fastq_directory, sample_name=sample_name, direction=direction
+        )
+        == expected_output_path
+    )
